@@ -4,15 +4,15 @@ description: Flutter BLoC example
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   meta: ^1.3.0
   equatable: ^2.0.0
   http: ^0.13.3
-  flutter_bloc: ^7.0.0
+  flutter_bloc: ^7.0.1
   rxdart: ^0.27.0
-  flutter_webview_plugin: ^0.3.11
+  flutter_webview_plugin: ^0.4.0
   easy_localization: ^3.0.0
   flutter:
     sdk: flutter
@@ -22,7 +22,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^5.0.7
+  mockito: ^5.0.9
+  bloc_test: ^8.0.1
 
 flutter:
   uses-material-design: true
