@@ -7,7 +7,7 @@ import 'package:flutter_bloc_lyrics/resources/langs/strings.dart';
 import 'package:flutter_bloc_lyrics/widgets/buttons.dart';
 
 class SongAddForm extends StatefulWidget {
-  final SongBase song;
+  final SongBase? song;
 
   SongAddForm({this.song});
 
@@ -19,13 +19,13 @@ class SongAddForm extends StatefulWidget {
 
 class SongAddState extends State<SongAddForm> {
   static final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
-  String _lyrics;
-  String _artist;
-  String _title;
+  String? _lyrics;
+  String? _artist;
+  String? _title;
 
-  SongAddEditBloc _songAddEditBloc;
+  late SongAddEditBloc _songAddEditBloc;
 
-  final SongBase _song;
+  final SongBase? _song;
 
   SongAddState(this._song);
 
@@ -38,63 +38,67 @@ class SongAddState extends State<SongAddForm> {
   @override
   Widget build(BuildContext context) {
     return BlocListener<SongAddEditBloc, SongAddEditState>(
-        bloc: _songAddEditBloc,
-        listener: (context, state) {
-          if (state is AddSongStateSuccess || state is EditSongStateSuccess) {
-            Navigator.pop(context);
-          }
-        },
-        child: Padding(
-            padding: EdgeInsets.symmetric(horizontal: 16.0),
-            child: Form(
-                key: _formKey,
-                child: Column(
-                  children: <Widget>[
-                    TextFormField(
-                      initialValue: _song != null ? _song.title : "",
-                      decoration: InputDecoration(hintText: S.TITLE.tr()),
-                      onSaved: (value) => _title = value,
-                      validator: (val) {
-                        return val.trim().isEmpty ? S.EMPTY_TITLE : null;
+      bloc: _songAddEditBloc,
+      listener: (context, state) {
+        if (state is AddSongStateSuccess || state is EditSongStateSuccess) {
+          Navigator.pop(context);
+        }
+      },
+      child: Padding(
+        padding: EdgeInsets.symmetric(horizontal: 16.0),
+        child: Form(
+          key: _formKey,
+          child: Column(
+            children: <Widget>[
+              TextFormField(
+                initialValue: _song?.title ?? "",
+                decoration: InputDecoration(hintText: S.TITLE.tr()),
+                onSaved: (value) => _title = value,
+                validator: (val) {
+                  return val?.trim().isEmpty == true ? S.EMPTY_TITLE.tr() : null;
+                },
+              ),
+              TextFormField(
+                initialValue: _song?.artist ?? "",
+                decoration: InputDecoration(hintText: S.ARTIST.tr()),
+                onSaved: (value) => _artist = value,
+                validator: (val) {
+                  return val?.trim().isEmpty == true ? S.EMPTY_ARTIST.tr() : null;
+                },
+              ),
+              TextFormField(
+                initialValue: _song?.lyrics ?? "",
+                decoration: InputDecoration(hintText: S.LYRICS.tr()),
+                onSaved: (value) => _lyrics = value,
+                validator: (val) {
+                  return val?.trim().isEmpty == true ? S.EMPTY_LYRICS.tr() : null;
+                },
+                minLines: 5,
+                maxLines: 20,
+              ),
+              Spacer(),
+              Padding(
+                  padding: EdgeInsets.only(bottom: 16.0),
+                  child: getBaseButton(
+                      onPressed: () {
+                        if (_formKey.currentState?.validate() == true) {
+                          _formKey.currentState?.save();
+                          SongBase updatedSong = SongBase(
+                            id: _song?.id,
+                            title: _title ?? "",
+                            lyrics: _lyrics ?? "",
+                            artist: _artist ?? "",
+                          );
+                          _songAddEditBloc.add(_song == null
+                              ? AddSong(song: updatedSong)
+                              : EditSong(song: updatedSong));
+                        }
                       },
-                    ),
-                    TextFormField(
-                      initialValue: _song != null ? _song.artist : "",
-                      decoration: InputDecoration(hintText: S.ARTIST.tr()),
-                      onSaved: (value) => _artist = value,
-                      validator: (val) {
-                        return val.trim().isEmpty ? S.EMPTY_ARTIST : null;
-                      },
-                    ),
-                    TextFormField(
-                      initialValue: _song != null ? _song.lyrics : "",
-                      decoration: InputDecoration(hintText: S.LYRICS.tr()),
-                      onSaved: (value) => _lyrics = value,
-                      validator: (val) {
-                        return val.trim().isEmpty ? S.EMPTY_LYRICS : null;
-                      },
-                      minLines: 5,
-                      maxLines: 20,
-                    ),
-                    Spacer(),
-                    Padding(
-                        padding: EdgeInsets.only(bottom: 16.0),
-                        child: getBaseButton(
-                            onPressed: () {
-                              if (_formKey.currentState.validate()) {
-                                _formKey.currentState.save();
-                                SongBase updatedSong = SongBase(
-                                    id: _song == null ? null : _song.id,
-                                    title: _title,
-                                    lyrics: _lyrics,
-                                    artist: _artist);
-                                _songAddEditBloc.add(_song == null
-                                    ? AddSong(song: updatedSong)
-                                    : EditSong(song: updatedSong));
-                              }
-                            },
-                            text: (_song != null ? S.EDIT : S.ADD_SONG).tr()))
-                  ],
-                ))));
+                      text: (_song != null ? S.EDIT : S.ADD_SONG).tr()))
+            ],
+          ),
+        ),
+      ),
+    );
   }
 }
