@@ -1,16 +1,21 @@
 import 'package:flutter/material.dart';
 
-SizedBox getBaseButton({String text, VoidCallback onPressed}) {
+SizedBox getBaseButton({
+  required String text,
+  required VoidCallback onPressed,
+}) {
   return SizedBox(
-      height: 48.0,
-      width: double.infinity,
-      child: RaisedButton(
-        child: Text(text),
-        onPressed: onPressed,
-        textColor: Colors.white,
-        color: Colors.blue,
+    height: 48.0,
+    width: double.infinity,
+    child: ElevatedButton(
+      onPressed: onPressed,
+      style: ElevatedButton.styleFrom(
+        primary: Colors.blue,
         shape: new RoundedRectangleBorder(
           borderRadius: new BorderRadius.circular(30.0),
         ),
-      ));
+      ),
+      child: Text(text),
+    ),
+  );
 }
