@@ -1,1 +1,1 @@
-const int DEFAULT_SEARCH_DEBOUNCE = 500;
+const DEFAULT_SEARCH_DEBOUNCE = 500;
