@@ -7,19 +7,21 @@ class SearchScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-          title: Text("Lyrics App"),
-        ),
-        body: Column(
-          children: <Widget>[SearchBar(), SongsSearchList()],
-        ),
-        floatingActionButton: FloatingActionButton(
-          onPressed: () {
-            Navigator.push(context,
-                MaterialPageRoute(builder: (context) => SongAddScreen()));
-          },
-          child: Icon(Icons.add),
-        ),
+      appBar: AppBar(
+        title: Text("Lyrics App"),
+      ),
+      body: Column(
+        children: <Widget>[SearchBar(), SongsSearchList()],
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          Navigator.push(
+            context,
+            MaterialPageRoute(builder: (context) => SongAddScreen()),
+          );
+        },
+        child: Icon(Icons.add),
+      ),
     );
   }
 }
