@@ -3,7 +3,7 @@ import 'package:flutter_bloc_lyrics/model/api/song_result.dart';
 class SearchResult {
   final SearchItems searchItems;
 
-  const SearchResult({this.searchItems});
+  const SearchResult({required this.searchItems});
 
   static SearchResult fromJson(dynamic json) {
     return SearchResult(searchItems: SearchItems.fromJson(json["response"]));
@@ -13,7 +13,7 @@ class SearchResult {
 class SearchItems {
   final List<SongSearchResult> songs;
 
-  const SearchItems({this.songs});
+  const SearchItems({required this.songs});
 
   static SearchItems fromJson(Map<String, dynamic> json) {
     final items = (json['hits'] as List<dynamic>)
