@@ -1,7 +1,7 @@
 class SearchResultError {
   final String message;
 
-  const SearchResultError({this.message});
+  const SearchResultError({required this.message});
 
   static SearchResultError fromJson(dynamic json) {
     return SearchResultError(message: json['message'] as String);
@@ -11,7 +11,7 @@ class SearchResultError {
 class MetaResponse {
   final SearchResultError searchResultError;
 
-  const MetaResponse({this.searchResultError});
+  const MetaResponse({required this.searchResultError});
 
   static MetaResponse fromJson(dynamic json) {
     return MetaResponse(
