@@ -1,9 +1,7 @@
-import 'package:flutter_bloc_lyrics/model/song_base.dart';
-
 class Artist {
   final String name;
 
-  const Artist({this.name});
+  const Artist({required this.name});
 
   static Artist fromJson(dynamic json) {
     return Artist(
